requirements.txt
streamlit
plotly
pandas
python-dotenv
requests